backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
requests==2.31.0
groq==1.0.0
PyJWT==2.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
firebase-admin==6.4.0
bcrypt==4.1.2
marshmallow==3.20.1
python-dateutil==2.8.2
Pillow>=10.2.0
validators==0.22.0
cryptography==46.0.4
beautifulsoup4==4.12.2
pymongo==4.6.1
numpy==1.24.3
scipy==1.11.4
textblob==0.17.1
scikit-learn==1.3.2
numpy>=1.26.0
scipy>=1.11.4
textblob==0.17.1
scikit-learn>=1.3.2
